LayoutTests/accessibility/self-referencing-aria-labelledby-expected.txt

X ?          Y Z
This tests that self-referencing aria-labelledby is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test.description is 'AXDescription: X Y Z'
PASS test.description is 'AXDescription: X Y Z'
PASS test.description is 'AXDescription: X Y Z'
PASS test.description is 'AXDescription: X Y Z'
PASS test.description is 'AXDescription: X Y Z'
PASS test.description is 'AXDescription: X Y Z'
PASS test.description is 'AXDescription: X Y Z'
PASS successfullyParsed is true

TEST COMPLETE
